Add Prime Factors tests, change IScalarProblemSolver.SolveProblem method to accept long instead of int

+++ Euler.Solver/IScalarProblemSolver.cs
@@ -2,6 +2,6 @@
 {
     public interface IScalarProblemSolver
     {
-        long SolveProblem(int input);
+        long SolveProblem(long input);
     }
 }
